Fixed some PHPCS issues, 7
