+++ src/php/CrdmBasic/Customizer/class-background.php
@@ -1,6 +1,6 @@
 <?php declare( strict_types=1 );
 
-namespace Crdm\Customizer;
+namespace CrdmBasic\Customizer;
 
 use Kirki;
 
@@ -39,7 +39,7 @@ class Background {
 				'section'   => $this->section_id,
 				'default'   => [
 					'background-color'      => '#f7f3e2',
-					'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_background.png',
+					'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_background.png',
 					'background-repeat'     => 'repeat',
 					'background-position'   => 'left top',
 					'background-size'       => '300px auto',
@@ -64,7 +64,7 @@ class Background {
 				'section'     => $this->section_id,
 				'default'     => [
 					'background-color'      => 'rgba(255, 255, 255, 0)',
-					'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_header_background.png',
+					'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_header_background.png',
 					'background-repeat'     => 'no-repeat',
 					'background-position'   => 'right top',
 					'background-size'       => '376px auto',
@@ -89,7 +89,7 @@ class Background {
 				'section'     => $this->section_id,
 				'default'     => [
 					'background-color'      => 'rgba(255, 255, 255, 0)',
-					'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_header_foreground.png',
+					'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_header_foreground.png',
 					'background-repeat'     => 'no-repeat',
 					'background-position'   => 'right bottom',
 					'background-size'       => '100% auto',
@@ -114,7 +114,7 @@ class Background {
 				'section'     => $this->section_id,
 				'default'     => [
 					'background-color'      => 'rgba(255, 255, 255, 0)',
-					'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_grass.png',
+					'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_grass.png',
 					'background-repeat'     => 'repeat-x',
 					'background-position'   => 'left bottom',
 					'background-size'       => 'auto 100%',
