+++ src/php/CrdmBasic/Customizer/class-border-radius.php
@@ -1,6 +1,6 @@
 <?php declare( strict_types=1 );
 
-namespace Crdm\Customizer;
+namespace CrdmBasic\Customizer;
 
 use Kirki;
