+++ src/php/CrdmBasic/Customizer/class-color-variant.php
@@ -1,6 +1,6 @@
 <?php declare( strict_types=1 );
 
-namespace Crdm\Customizer;
+namespace CrdmBasic\Customizer;
 
 use Kirki;
 
@@ -40,19 +40,19 @@ class Color_Variant {
 				'section'     => $this->section_id,
 				'default'     => 'light',
 				'choices'     => [
-					'light' => CRDM_BASIC_TEMPLATE_URL . 'admin/light.png',
-					'dark'  => CRDM_BASIC_TEMPLATE_URL . 'admin/dark.png',
+					'light' => CRDMBASIC_TEMPLATE_URL . 'admin/light.png',
+					'dark'  => CRDMBASIC_TEMPLATE_URL . 'admin/dark.png',
 				],
 				'preset'      => [
 					'light' => [
 						'settings' => [
 							'webBg'                 => [
 								'background-color' => 'rgba(255, 255, 255, 0)',
-								'background-image' => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_background.png',
+								'background-image' => CRDMBASIC_TEMPLATE_URL . 'frontend/light_background.png',
 							],
 							'headerBg1'             => [
 								'background-color'      => 'rgba(255, 255, 255, 0)',
-								'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_header_background.png',
+								'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_header_background.png',
 								'background-repeat'     => 'no-repeat',
 								'background-position'   => 'right top',
 								'background-size'       => '376px auto',
@@ -60,7 +60,7 @@ class Color_Variant {
 							],
 							'headerBg2'             => [
 								'background-color'      => 'rgba(255, 255, 255, 0)',
-								'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_header_foreground.png',
+								'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_header_foreground.png',
 								'background-repeat'     => 'no-repeat',
 								'background-position'   => 'right bottom',
 								'background-size'       => '100% auto',
@@ -68,7 +68,7 @@ class Color_Variant {
 							],
 							'headerBg3'             => [
 								'background-color'      => 'rgba(255, 255, 255, 0)',
-								'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/light_grass.png',
+								'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/light_grass.png',
 								'background-repeat'     => 'repeat-x',
 								'background-position'   => 'left bottom',
 								'background-size'       => 'auto 100%',
@@ -138,7 +138,7 @@ class Color_Variant {
 							],
 							'headerBg2'             => [
 								'background-color'      => 'rgba(255, 255, 255, 0)',
-								'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/dark_header_foreground.png',
+								'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/dark_header_foreground.png',
 								'background-repeat'     => 'no-repeat',
 								'background-position'   => 'right bottom',
 								'background-size'       => '87% auto',
@@ -146,7 +146,7 @@ class Color_Variant {
 							],
 							'headerBg3'             => [
 								'background-color'      => 'rgba(255, 255, 255, 0)',
-								'background-image'      => CRDM_BASIC_TEMPLATE_URL . 'frontend/dark_pavement.png',
+								'background-image'      => CRDMBASIC_TEMPLATE_URL . 'frontend/dark_pavement.png',
 								'background-repeat'     => 'repeat-x',
 								'background-position'   => 'left bottom',
 								'background-size'       => '24px 10px',
